third_party/WebKit/LayoutTests/external/wpt/payment-request/allowpaymentrequest/setting-allowpaymentrequest-timing.https.sub-expected.txt

This is a testharness.js-based test.
FAIL PaymentRequest setting allowpaymentrequest after document creation, before response assert_equals: expected "Success" but got "Exception"
Harness: the test ran to completion.
